+++ infopark_reactor_migrations/lib/reactor/cm/xml_multi_request_error.rb
@@ -1,0 +1,10 @@
+# -*- encoding : utf-8 -*-
+require 'reactor/cm/xml_request_error'
+
+module Reactor
+  module Cm
+    class XmlMultiRequestError < XmlRequestError
+
+    end
+  end
+end