+++ reactor_test/use-config.rb
@@ -5,6 +5,20 @@ require 'ostruct'
 require 'pp'
 
 $CONFIGS = {
+  'ruby2.1.2+Rails4.0.9+infopark_fiona_connector-beta' => {
+    '.ruby-version' => '2.1.2',
+    'Gemfile' => <<-EOGEMFILE
+source "https://rubygems.org"
+gem "infopark_rails_connector_meta", :path=>"../infopark_rails_connector_meta"
+gem "infopark_reactor_migrations", :path=>"../infopark_reactor_migrations"
+gem "infopark_reactor", :path=>"../infopark_reactor"
+gem "rspec-rails", "~> 2.0"
+gem "infopark_fiona_connector", :git => "git@github.com:infopark/fiona_connector.git", :branch => 'dev'
+gem "mysql2"
+gem "nokogiri", "< 1.6.0"
+gem "rails", "4.0.9"
+    EOGEMFILE
+  },
   'ruby2.1.2+Rails4.0.8+infopark_fiona_connector-beta' => {
     '.ruby-version' => '2.1.2',
     'Gemfile' => <<-EOGEMFILE
@@ -48,6 +62,21 @@ gem "rails", "3.2.19"
 gem "nokogiri", "< 1.6.0"
     EOGEMFILE
   },
+  'ruby1.9.3-p547+Rails3.2.20+infopark_fiona_connector-6.9.4' => {
+    '.ruby-version' => '1.9.3-p547',
+    'Gemfile' => <<-EOGEMFILE
+source "https://rubygems.org"
+gem "infopark_rails_connector_meta", :path=>"../infopark_rails_connector_meta"
+gem "infopark_reactor_migrations", :path=>"../infopark_reactor_migrations"
+gem "infopark_reactor", :path=>"../infopark_reactor"
+gem "rspec-rails", "~> 2.0"
+gem "infopark_rails_connector", "6.9.4"
+gem "infopark_fiona_connector", "6.9.4"
+gem "mysql2"
+gem "rails", "3.2.20"
+gem "nokogiri", "< 1.6.0"
+    EOGEMFILE
+  },
   'ruby1.9.3-p547+Rails3.2.19+infopark_fiona_connector-6.9.1.3.22208381' => {
     '.ruby-version' => '1.9.3-p547',
     'Gemfile' => <<-EOGEMFILE
@@ -65,7 +94,7 @@ gem "nokogiri", "< 1.6.0"
   }
 }
 
-$DEFUALT = 'ruby2.1.2+Rails4.0.8+infopark_fiona_connector-beta'
+$DEFUALT = 'ruby2.1.2+Rails4.0.9+infopark_fiona_connector-beta'
 
 $DAMN_YOU_RBENV = lambda {|command| "bash -c 'eval \"$(rbenv init -)\"; rbenv shell $(cat .ruby-version); #{command}'" }